Refactor - move account icon to Navigation component
